+++ Cargo.toml
@@ -123,6 +123,9 @@ stream-broadcast = [
   "dep:libc",
   "dep:glow",
   "dep:khronos-egl",
+  "x11rb/image",
+  "x11rb/randr",
+  "x11rb/xfixes",
   "dep:wasapi",
   "dep:windows",
   "dep:block2",
